Unified CastCache implementation
