+++ src/coreclr/nativeaot/Runtime.Base/src/System/Runtime/TypeCast.cs
@@ -107,7 +107,7 @@ namespace System.Runtime
                 // parameters are compatible.
 
                 // NOTE: using general assignable path for the cache because of the cost of the variance checks
-                if (CastCache.AreTypesAssignableInternal(pObjType, pTargetType, AssignmentVariation.BoxedSource, null))
+                if (AreTypesAssignableInternal(pObjType, pTargetType, AssignmentVariation.BoxedSource, null))
                     return obj;
                 return null;
             }
@@ -207,7 +207,7 @@ namespace System.Runtime
                 }
             }
 
-            if (CastCache.AreTypesAssignableInternal(pObjType->RelatedParameterType, pTargetType->RelatedParameterType,
+            if (AreTypesAssignableInternal(pObjType->RelatedParameterType, pTargetType->RelatedParameterType,
                 AssignmentVariation.AllowSizeEquivalence, null))
             {
                 return obj;
@@ -246,7 +246,7 @@ namespace System.Runtime
 
             MethodTable* pObjType = obj.GetMethodTable();
 
-            if (CastCache.AreTypesAssignableInternal_SourceNotTarget_BoxedSource(pObjType, pTargetType, null))
+            if (AreTypesAssignableInternal_SourceNotTarget_BoxedSource(pObjType, pTargetType, null))
                 return obj;
 
             // If object type implements IDynamicInterfaceCastable then there's one more way to check whether it implements
@@ -441,7 +441,7 @@ namespace System.Runtime
                         //   class Foo : ICovariant<Bar> is ICovariant<IBar>
                         //   class Foo : ICovariant<IBar> is ICovariant<Object>
 
-                        if (!CastCache.AreTypesAssignableInternal(pSourceArgType, pTargetArgType, AssignmentVariation.Normal, pVisited))
+                        if (!AreTypesAssignableInternal(pSourceArgType, pTargetArgType, AssignmentVariation.Normal, pVisited))
                             return false;
 
                         break;
@@ -456,7 +456,7 @@ namespace System.Runtime
 
                         // This call is just like the call for Covariance above except true is passed
                         // to the fAllowSizeEquivalence parameter to allow the int/uint matching to work
-                        if (!CastCache.AreTypesAssignableInternal(pSourceArgType, pTargetArgType, AssignmentVariation.AllowSizeEquivalence, pVisited))
+                        if (!AreTypesAssignableInternal(pSourceArgType, pTargetArgType, AssignmentVariation.AllowSizeEquivalence, pVisited))
                             return false;
 
                         break;
@@ -471,7 +471,7 @@ namespace System.Runtime
                         //   class Foo : IContravariant<IBar> is IContravariant<Bar>
                         //   class Foo : IContravariant<Object> is IContravariant<IBar>
 
-                        if (!CastCache.AreTypesAssignableInternal(pTargetArgType, pSourceArgType, AssignmentVariation.Normal, pVisited))
+                        if (!AreTypesAssignableInternal(pTargetArgType, pSourceArgType, AssignmentVariation.Normal, pVisited))
                             return false;
 
                         break;
@@ -513,7 +513,7 @@ namespace System.Runtime
                 return AreTypesEquivalent(pSourceType, pNullableType);
             }
 
-            return CastCache.AreTypesAssignableInternal(pSourceType, pTargetType, AssignmentVariation.BoxedSource, null);
+            return AreTypesAssignableInternal(pSourceType, pTargetType, AssignmentVariation.BoxedSource, null);
         }
 
         // Internally callable version of the export method above. Has two additional flags:
@@ -521,7 +521,7 @@ namespace System.Runtime
         //                            compatible with Object, ValueType and Enum (if applicable)
         //  fAllowSizeEquivalence   : allow identically sized integral types and enums to be considered
         //                            equivalent (currently used only for array element types)
-        internal static unsafe bool AreTypesAssignableInternal(MethodTable* pSourceType, MethodTable* pTargetType, AssignmentVariation variation, EETypePairList* pVisited)
+        internal static unsafe bool AreTypesAssignableInternalUncached(MethodTable* pSourceType, MethodTable* pTargetType, AssignmentVariation variation, EETypePairList* pVisited)
         {
             bool fBoxedSource = ((variation & AssignmentVariation.BoxedSource) == AssignmentVariation.BoxedSource);
             bool fAllowSizeEquivalence = ((variation & AssignmentVariation.AllowSizeEquivalence) == AssignmentVariation.AllowSizeEquivalence);
@@ -585,7 +585,7 @@ namespace System.Runtime
                         // Note that using AreTypesAssignableInternal with AssignmentVariation.AllowSizeEquivalence
                         // here handles array covariance as well as IFoo[] -> Foo[] etc.  We are not using
                         // AssignmentVariation.BoxedSource because int[] is not assignable to object[].
-                        return CastCache.AreTypesAssignableInternal(pSourceType->RelatedParameterType,
+                        return AreTypesAssignableInternal(pSourceType->RelatedParameterType,
                             pTargetType->RelatedParameterType, AssignmentVariation.AllowSizeEquivalence, pVisited);
                     }
                 }
@@ -663,7 +663,7 @@ namespace System.Runtime
 
             MethodTable* pObjType = obj.GetMethodTable();
 
-            if (CastCache.AreTypesAssignableInternal_SourceNotTarget_BoxedSource(pObjType, pTargetType, null))
+            if (AreTypesAssignableInternal_SourceNotTarget_BoxedSource(pObjType, pTargetType, null))
                 return obj;
 
             // If object type implements IDynamicInterfaceCastable then there's one more way to check whether it implements
@@ -690,7 +690,7 @@ namespace System.Runtime
             Debug.Assert(array.GetMethodTable()->IsArray, "first argument must be an array");
 
             MethodTable* arrayElemType = array.GetMethodTable()->RelatedParameterType;
-            if (CastCache.AreTypesAssignableInternal(obj.GetMethodTable(), arrayElemType, AssignmentVariation.BoxedSource, null))
+            if (AreTypesAssignableInternal(obj.GetMethodTable(), arrayElemType, AssignmentVariation.BoxedSource, null))
                 return;
 
             // If object type implements IDynamicInterfaceCastable then there's one more way to check whether it implements
@@ -795,7 +795,7 @@ namespace System.Runtime
         [MethodImpl(MethodImplOptions.NoInlining)]
         private static unsafe void StelemRef_Helper(ref object element, MethodTable* elementType, object obj)
         {
-            if (CastCache.AreTypesAssignableInternal(obj.GetMethodTable(), elementType, AssignmentVariation.BoxedSource, null))
+            if (AreTypesAssignableInternal(obj.GetMethodTable(), elementType, AssignmentVariation.BoxedSource, null))
             {
                 InternalCalls.RhpAssignRef(ref element, obj);
             }
@@ -1034,495 +1034,63 @@ namespace System.Runtime
             }
         }
 
-        // source type + target type + assignment variation -> true/false
-        [System.Runtime.CompilerServices.EagerStaticClassConstructionAttribute]
-        internal static unsafe class CastCache
+        public static unsafe bool AreTypesAssignableInternal(MethodTable* pSourceType, MethodTable* pTargetType, AssignmentVariation variation, EETypePairList* pVisited)
         {
-#if DEBUG
-            private const int INITIAL_CACHE_SIZE = 8;    // MUST BE A POWER OF TWO
-            private const int MAXIMUM_CACHE_SIZE = 512;  // make this lower than release to make it easier to reach this in tests.
-#else
-            private const int INITIAL_CACHE_SIZE = 128;  // MUST BE A POWER OF TWO
-            private const int MAXIMUM_CACHE_SIZE = 4096; // 4096 * sizeof(CastCacheEntry) is 98304 bytes on 64bit. We will rarely need this much though.
-#endif
-            private const int BUCKET_SIZE = 8;
-
-            private const int VERSION_NUM_SIZE = 29;
-            private const uint VERSION_NUM_MASK = (1 << VERSION_NUM_SIZE) - 1;
-
-            // A trivial 2-elements table used for "flushing" the cache. Nothing is ever stored in this table.
-            // It is required that we are able to allocate this.
-            private static int[] s_sentinelTable = CreateCastCache(2, throwOnFail: true)!;
-
-            // when flushing, remember the last size.
-            private static int s_lastFlushSize = INITIAL_CACHE_SIZE;
-
-            // The actual storage.
-            private static int[] s_table = CreateCastCache(INITIAL_CACHE_SIZE) ?? s_sentinelTable;
-
-            // TODO: VS
-            // [DebuggerDisplay("Source = {_source}; Target = {_targetAndResult & ~1}; Result = {_targetAndResult & 1}; VersionNum = {_version & ((1 << 29) - 1)}; Distance = {_version >> 29};")]
-
-            [StructLayout(LayoutKind.Sequential)]
-            private struct CastCacheEntry
-            {
-                // version has the following structure:
-                // [ distance:3bit |  versionNum:29bit ]
-                //
-                // distance is how many iterations the entry is from it ideal position.
-                // we use that for preemption.
-                //
-                // versionNum is a monotonically increasing numerical tag.
-                // Writer "claims" entry by atomically incrementing the tag. Thus odd number indicates an entry in progress.
-                // Upon completion of adding an entry the tag is incremented again making it even. Even number indicates a complete entry.
-                //
-                // Readers will read the version twice before and after retrieving the entry.
-                // To have a usable entry both reads must yield the same even version.
-                //
-                internal uint _version;
-                internal nuint _source;
-                // pointers have unused lower bits due to alignment, we use one for the result
-                internal nuint _targetAndResult;
-
-                internal void SetEntry(nuint source, nuint target, bool result)
-                {
-                    _source = source;
-                    _targetAndResult = target | (nuint)(result ? 1: 0);
-                }
-            }
-            // this is to avoid dependency on BitOperations
-#if TARGET_64BIT
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static ulong RotateLeft(ulong value, int offset)
-                => (value << offset) | (value >> (64 - offset));
-#else
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static uint RotateLeft(uint value, int offset)
-                => (value << offset) | (value >> (32 - offset));
-#endif
-            private struct VolatileUInt32 { public volatile uint Value; }
-
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static uint VolatileRead(ref uint location) =>
-                Unsafe.As<uint, VolatileUInt32>(ref location).Value;
-
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static void VolatileWrite(ref uint location, uint value)
-            {
-                Unsafe.As<uint, VolatileUInt32>(ref location).Value = value;
-            }
-
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static uint InterlockedCompareExchange(ref uint location1, uint value, uint comparand) =>
-                (uint)Interlocked.CompareExchange(ref Unsafe.As<uint, int>(ref location1), (int)value, (int)comparand);
-
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static int KeyToBucket(ref int tableData, nuint source, nuint target)
-            {
-                // upper bits of addresses do not vary much, so to reduce loss due to cancelling out,
-                // we do `rotl(source, <half-size>) ^ target` for mixing inputs.
-                // then we use fibonacci hashing to reduce the value to desired size.
-
-                int hashShift = HashShift(ref tableData);
-#if TARGET_64BIT
-                ulong hash = RotateLeft((ulong)source, 32) ^ (ulong)target;
-                return (int)((hash * 11400714819323198485ul) >> hashShift);
-#else
-                uint hash = RotateLeft((uint)source, 16) ^ (uint)target;
-                return (int)((hash * 2654435769u) >> hashShift);
-#endif
-            }
-
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static ref int TableData(int[] table)
-            {
-                // element 0 is used for embedded aux data
-                // return ref MemoryMarshal.GetArrayDataReference(table);
-                return ref Unsafe.As<byte, int>(ref Unsafe.AddByteOffset(ref table.GetRawData(), (nint)sizeof(nint)));
-            }
-
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static ref CastCacheEntry Element(ref int tableData, int index)
-            {
-                // element 0 is used for embedded aux data, skip it
-                return ref Unsafe.Add(ref Unsafe.As<int, CastCacheEntry>(ref tableData), index + 1);
-            }
-
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static ref int HashShift(ref int tableData)
-            {
-                return ref tableData;
-            }
-
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static ref uint VictimCounter(ref int tableData)
-            {
-                return ref Unsafe.As<int, uint>(ref Unsafe.Add(ref tableData, 2));
-            }
-
-            // TableMask is "size - 1"
-            // we need that more often that we need size
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static ref int TableMask(ref int tableData)
-            {
-                return ref Unsafe.Add(ref tableData, 1);
-            }
-
-            private enum CastResult
-            {
-                CannotCast = 0,
-                CanCast = 1,
-                MaybeCast = 2
-            }
-
-            private static int[]? CreateCastCache(int size, bool throwOnFail = false)
-            {
-                // size must be positive
-                Debug.Assert(size > 1);
-                // size must be a power of two
-                Debug.Assert((size & (size - 1)) == 0);
-
-                int[]? table = null;
-                try
-                {
-                    table = new int[(size + 1) * sizeof(CastCacheEntry) / sizeof(int)];
-                }
-                catch (OutOfMemoryException)
-                {
-                }
-
-                if (table == null)
-                {
-                    size = INITIAL_CACHE_SIZE;
-                    try
-                    {
-                        table = new int[(size + 1) * sizeof(CastCacheEntry) / sizeof(int)];
-                    }
-                    catch (OutOfMemoryException)
-                    {
-                    }
-                }
-
-                if (table == null)
-                {
-                    // we do not OOM in casts, just return null meaning we can't resize.
-                    return table;
-                }
-
-                ref int tableData = ref TableData(table);
-
-                // set the table mask. we need it often, do not want to compute each time.
-                TableMask(ref tableData) = size - 1;
-
-                // Fibonacci hash reduces the value into desired range by shifting right by the number of leading zeroes in 'size-1'
-                int bitCnt = 0;
-
-                // software fallback for BitOperations.Log2. Size should be a small number anyways.
-                int mask = size - 1;
-                while (mask != 0)
-                {
-                    mask >>= 1;
-                    bitCnt++;
-                }
-
-#if TARGET_64BIT
-                HashShift(ref tableData) = (byte)(64 - bitCnt);
-#else
-                HashShift(ref tableData) = (byte)(32 - bitCnt);
-#endif
-
-                return table;
-            }
-
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            private static CastResult TryGet(nuint source, nuint target)
-            {
-                const int BUCKET_SIZE = 8;
-
-                // table is initialized and is not null.
-                ref int tableData = ref TableData(s_table);
-
-                int index = KeyToBucket(ref tableData, source, target);
-                for (int i = 0; i < BUCKET_SIZE;)
-                {
-                    ref CastCacheEntry pEntry = ref Element(ref tableData, index);
-
-                    // must read in this order: version -> [entry parts] -> version
-                    // if version is odd or changes, the entry is inconsistent and thus ignored
-                    uint version = VolatileRead(ref pEntry._version);
-                    nuint entrySource = pEntry._source;
-
-                    // mask the lower version bit to make it even.
-                    // This way we can check if version is odd or changing in just one compare.
-                    version &= unchecked((uint)~1);
-
-                    if (entrySource == source)
-                    {
-                        nuint entryTargetAndResult = pEntry._targetAndResult;
-                        // target never has its lower bit set.
-                        // a matching entryTargetAndResult would the have same bits, except for the lowest one, which is the result.
-                        entryTargetAndResult ^= target;
-                        if (entryTargetAndResult <= 1)
-                        {
-                            // make sure 'version' is loaded after 'source' and 'targetAndResults'
-                            //
-                            // We can either:
-                            // - use acquires for both _source and _targetAndResults or
-                            // - issue a load barrier before reading _version
-                            // benchmarks on available hardware show that use of a read barrier is cheaper.
-
-                            // TODO: VS
-                            // Interlocked.ReadMemoryBarrier();
-
-                            if (version != pEntry._version)
-                            {
-                                // oh, so close, the entry is in inconsistent state.
-                                // it is either changing or has changed while we were reading.
-                                // treat it as a miss.
-                                break;
-                            }
-
-                            return (CastResult)entryTargetAndResult;
-                        }
-                    }
-
-                    if (version == 0)
-                    {
-                        // the rest of the bucket is unclaimed, no point to search further
-                        break;
-                    }
-
-                    // quadratic reprobe
-                    i++;
-                    index = (index + i) & TableMask(ref tableData);
-                }
-                return CastResult.MaybeCast;
-            }
-
-            private static int CacheElementCount(ref int tableData)
-            {
-                return TableMask(ref tableData) + 1;
-            }
-
-            private static bool MaybeReplaceCacheWithLarger(int size)
-            {
-                int[]? newTable = CreateCastCache(size);
-                if (newTable == null)
-                {
-                    return false;
-                }
-
-                s_table = newTable;
+            // Important special case -- it breaks infinite recursion in CastCache itself!
+            if (pSourceType == pTargetType)
                 return true;
+
+            nuint sourceAndVariation = (nuint)pSourceType + (uint)variation;
+            CastResult result = CastCache.TryGet(sourceAndVariation, (nuint)(pTargetType));
+            if (result != CastResult.MaybeCast)
+            {
+                return result == CastResult.CanCast;
             }
 
-            private static bool TryGrow(ref int tableData)
-            {
-                int newSize = CacheElementCount(ref tableData) * 2;
-                if (newSize <= MAXIMUM_CACHE_SIZE)
-                {
-                    return MaybeReplaceCacheWithLarger(newSize);
-                }
+            return CacheMiss(pSourceType, pTargetType, variation, pVisited);
+        }
 
+        // This method is an optimized and customized version of AreTypesAssignable that achieves better performance
+        // than AreTypesAssignableInternal through 2 significant changes
+        // 1. Removal of sourceType to targetType check (This property must be known before calling this function. At time
+        //    of writing, this is true as its is only used if sourceType is from an object, and targetType is an interface.)
+        // 2. Force inlining (This particular variant is only used in a small number of dispatch scenarios that are particularly
+        //    high in performance impact.)
+        // [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static unsafe bool AreTypesAssignableInternal_SourceNotTarget_BoxedSource(MethodTable* pSourceType, MethodTable* pTargetType, EETypePairList* pVisited)
+        {
+            Debug.Assert(pSourceType != pTargetType, "target is source");
+            nuint sourceAndVariation = (nuint)pSourceType + (int)AssignmentVariation.BoxedSource;
+            CastResult result = CastCache.TryGet(sourceAndVariation, (nuint)(pTargetType));
+            if (result != CastResult.MaybeCast)
+            {
+                return result == CastResult.CanCast;
+            }
+
+            return CacheMiss(pSourceType, pTargetType, AssignmentVariation.BoxedSource, pVisited);
+        }
+
+        private static unsafe bool CacheMiss(MethodTable* pSourceType, MethodTable* pTargetType, AssignmentVariation variation, EETypePairList* pVisited)
+        {
+            //
+            // First, check if we previously visited the input types pair, to avoid infinite recursions
+            //
+            if (EETypePairList.Exists(pVisited, pSourceType, pTargetType))
                 return false;
-            }
 
-            private static void TrySet(nuint source, nuint target, bool result)
-            {
-                int bucket;
-                ref int tableData = ref *(int*)0;
+            //
+            // Call into the type cast code to calculate the result
+            //
+            EETypePairList newList = new EETypePairList(pSourceType, pTargetType, pVisited);
+            bool result = TypeCast.AreTypesAssignableInternalUncached(pSourceType, pTargetType, variation, &newList);
 
-                do
-                {
-                    tableData = ref TableData(s_table);
-                    if (TableMask(ref tableData) == 1)
-                    {
-                        // 2-element table is used as a sentinel.
-                        // we did not allocate a real table yet or have flushed it.
-                        // try replacing the table, but do not insert anything.
-                        MaybeReplaceCacheWithLarger(s_lastFlushSize);
-                        return;
-                    }
+            //
+            // Update the cache
+            //
+            nuint sourceAndVariation = (nuint)pSourceType + (uint)variation;
+            CastCache.TrySet(sourceAndVariation, (nuint)pTargetType, result);
 
-                    bucket = KeyToBucket(ref tableData, source, target);
-                    int index = bucket;
-                    ref CastCacheEntry pEntry = ref Element(ref tableData, index);
-
-                    for (int i = 0; i < BUCKET_SIZE;)
-                    {
-                        // claim the entry if unused or is more distant than us from its origin.
-                        // Note - someone familiar with Robin Hood hashing will notice that
-                        //        we do the opposite - we are "robbing the poor".
-                        //        Robin Hood strategy improves average lookup in a lossles dictionary by reducing
-                        //        outliers via giving preference to more distant entries.
-                        //        What we have here is a lossy cache with outliers bounded by the bucket size.
-                        //        We improve average lookup by giving preference to the "richer" entries.
-                        //        If we used Robin Hood strategy we could eventually end up with all
-                        //        entries in the table being maximally "poor".
-
-                        // VolatileLoadWithoutBarrier is to ensure that the version cannot be re-fetched between here and CompareExchange.
-                        uint version = pEntry._version;
-
-                        // mask the lower version bit to make it even.
-                        // This way we will detect both if version is changing (odd) or has changed (even, but different).
-                        version &= unchecked((uint)~1);
-
-                        if ((version & VERSION_NUM_MASK) >= (VERSION_NUM_MASK - 2))
-                        {
-                            // If exactly VERSION_NUM_MASK updates happens between here and publishing, we may not recognize a race.
-                            // It is extremely unlikely, but to not worry about the possibility, lets not allow version to go this high and just get a new cache.
-                            // This will not happen often.
-                            FlushCurrentCache();
-                            return;
-                        }
-
-                        if (version == 0 || (version >> VERSION_NUM_SIZE) > i)
-                        {
-                            uint newVersion = ((uint)i << VERSION_NUM_SIZE) + (version & VERSION_NUM_MASK) + 1;
-                            uint versionOrig = InterlockedCompareExchange(ref pEntry._version, newVersion, version);
-                            if (versionOrig == version)
-                            {
-                                pEntry.SetEntry(source, target, result);
-
-                                // entry is in inconsistent state and cannot be read or written to until we
-                                // update the version, which is the last thing we do here
-                                VolatileWrite(ref pEntry._version, newVersion + 1);
-                                return;
-                            }
-                            // someone snatched the entry. try the next one in the bucket.
-                        }
-
-                        if (pEntry._source == source && ((pEntry._targetAndResult ^ target) <= 1))
-                        {
-                            // looks like we already have an entry for this.
-                            // duplicate entries are harmless, but a bit of a waste.
-                            return;
-                        }
-
-                        // quadratic reprobe
-                        i++;
-                        index += i;
-                        pEntry = ref Element(ref tableData, index & TableMask(ref tableData));
-                    }
-
-                    // bucket is full.
-                } while (TryGrow(ref tableData));
-
-                // reread tableData after TryGrow.
-                tableData = ref TableData(s_table);
-
-                if (TableMask(ref tableData) == 1)
-                {
-                    // do not insert into a sentinel.
-                    return;
-                }
-
-                // pick a victim somewhat randomly within a bucket
-                // NB: ++ is not interlocked. We are ok if we lose counts here. It is just a number that changes.
-                uint victimDistance = VictimCounter(ref tableData)++ & (BUCKET_SIZE - 1);
-                // position the victim in a quadratic reprobe bucket
-                uint victim = (victimDistance * victimDistance + victimDistance) / 2;
-
-                {
-                    ref CastCacheEntry pEntry = ref Element(ref tableData, (bucket + (int)victim) & TableMask(ref tableData));
-
-                    // VolatileLoadWithoutBarrier is to ensure that the version cannot be re-fetched between here and CompareExchange.
-                    uint version = pEntry._version;
-
-                    // mask the lower version bit to make it even.
-                    // This way we will detect both if version is changing (odd) or has changed (even, but different).
-                    version &= unchecked((uint)~1);
-
-                    if ((version & VERSION_NUM_MASK) >= (VERSION_NUM_MASK - 2))
-                    {
-                        // If exactly VERSION_NUM_MASK updates happens between here and publishing, we may not recognize a race.
-                        // It is extremely unlikely, but to not worry about the possibility, lets not allow version to go this high and just get a new cache.
-                        // This will not happen often.
-                        FlushCurrentCache();
-                        return;
-                    }
-
-                    uint newVersion = (victimDistance << VERSION_NUM_SIZE) + (version & VERSION_NUM_MASK) + 1;
-                    uint versionOrig = InterlockedCompareExchange(ref pEntry._version, newVersion, version);
-
-                    if (versionOrig == version)
-                    {
-                        pEntry.SetEntry(source, target, result);
-                        VolatileWrite(ref pEntry._version, newVersion + 1);
-                    }
-                }
-            }
-
-            private static void FlushCurrentCache()
-            {
-                ref int tableData = ref TableData(s_table);
-                int lastSize = CacheElementCount(ref tableData);
-                if (lastSize < INITIAL_CACHE_SIZE)
-                    lastSize = INITIAL_CACHE_SIZE;
-
-                s_lastFlushSize = lastSize;
-                // flushing is just replacing the table with a sentinel.
-                s_table = s_sentinelTable;
-            }
-
-            public static unsafe bool AreTypesAssignableInternal(MethodTable* pSourceType, MethodTable* pTargetType, AssignmentVariation variation, EETypePairList* pVisited)
-            {
-                // Important special case -- it breaks infinite recursion in CastCache itself!
-                if (pSourceType == pTargetType)
-                    return true;
-
-                nuint sourceAndVariation = (nuint)pSourceType + (uint)variation;
-                CastResult result = TryGet(sourceAndVariation, (nuint)(pTargetType));
-                if (result != CastResult.MaybeCast)
-                {
-                    return result == CastResult.CanCast;
-                }
-
-                return CacheMiss(pSourceType, pTargetType, variation, pVisited);
-            }
-
-            // This method is an optimized and customized version of AreTypesAssignable that achieves better performance
-            // than AreTypesAssignableInternal through 2 significant changes
-            // 1. Removal of sourceType to targetType check (This property must be known before calling this function. At time
-            //    of writing, this is true as its is only used if sourceType is from an object, and targetType is an interface.)
-            // 2. Force inlining (This particular variant is only used in a small number of dispatch scenarios that are particularly
-            //    high in performance impact.)
-            // [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public static unsafe bool AreTypesAssignableInternal_SourceNotTarget_BoxedSource(MethodTable* pSourceType, MethodTable* pTargetType, EETypePairList* pVisited)
-            {
-                Debug.Assert(pSourceType != pTargetType, "target is source");
-                nuint sourceAndVariation = (nuint)pSourceType + (int)AssignmentVariation.BoxedSource;
-                CastResult result = TryGet(sourceAndVariation, (nuint)(pTargetType));
-                if (result != CastResult.MaybeCast)
-                {
-                    return result == CastResult.CanCast;
-                }
-
-                return CacheMiss(pSourceType, pTargetType, AssignmentVariation.BoxedSource, pVisited);
-            }
-
-            private static unsafe bool CacheMiss(MethodTable* pSourceType, MethodTable* pTargetType, AssignmentVariation variation, EETypePairList* pVisited)
-            {
-                //
-                // First, check if we previously visited the input types pair, to avoid infinite recursions
-                //
-                if (EETypePairList.Exists(pVisited, pSourceType, pTargetType))
-                    return false;
-
-                //
-                // Call into the type cast code to calculate the result
-                //
-                EETypePairList newList = new EETypePairList(pSourceType, pTargetType, pVisited);
-                bool result = TypeCast.AreTypesAssignableInternal(pSourceType, pTargetType, variation, &newList);
-
-                //
-                // Update the cache
-                //
-                nuint sourceAndVariation = (nuint)pSourceType + (uint)variation;
-                TrySet(sourceAndVariation, (nuint)pTargetType, result);
-
-                return result;
-            }
+            return result;
         }
     }
 }
